Add files via upload
This is the first program I wrote with C++. It isn't something special for now but I'm just getting started. The program takes input from the user while the artificial opponent generates a random value (rock, paper, scissors), after that, both values get compared and the winner is proclaimed. Rounds last as long as the user wants and another match can be run after that.